+++ juliet/__init__.py
@@ -1,6 +1,6 @@
 __all__ = ['fit', 'utils', 'plots']
 
-__version__ = "2.0.25"
+__version__ = "2.0.26"
 
 from .fit import *
 from .utils import *